// src/api/models/SettingSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SettingSchema = {
    userId?: string;
    adminId?: string;
    companySize?: string;
    industry?: string;
    capital?: number;
    financialKnowledge?: string;
    fiscalYearStartYear?: number;
    fiscalYearStartMonth?: number;
    lastAdminCommentDate?: string;
};